requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0 
ultralytics==8.1.34 
torch
pillow